# docs/openapi/critical-stable.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DashClaw Critical Stable API",
    "version": "4.45.0",
    "description": "Generated OpenAPI spec for critical stable DashClaw endpoints. Source: app/api/**/route.js"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/guard/decisions": {
      "get": {
        "operationId": "get_api_guard_decisions",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or request error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "GET /api/guard/decisions",
        "tags": [
          "guard"
        ],
        "x-api-maturity": "stable"
      }
    },
...
  }
}